auth flow

+++ app/src/main/java/com/normalnywork/tundramarket/data/remote/api/mappers/LocationMapper.kt
@@ -4,6 +4,6 @@ import com.normalnywork.tundramarket.data.remote.api.proto.ProtoLocation
 import com.normalnywork.tundramarket.domain.entities.Location
 
 fun ProtoLocation.toDomain() = Location(
-    latitude = latitude.toFloat(),
-    longitude = longitude.toFloat(),
+    latitude = latitude,
+    longitude = longitude,
 )
